+++ src/Language/Hasmtlib/Counting.hs
@@ -3,6 +3,7 @@ module Language.Hasmtlib.Counting where
 import Prelude hiding (not, (&&), (||), or)
 import Language.Hasmtlib.Internal.Expr.Num ()
 import Language.Hasmtlib.Internal.Expr
+import Language.Hasmtlib.Type.SMTSort
 import Language.Hasmtlib.Equatable
 import Language.Hasmtlib.Orderable
 import Language.Hasmtlib.Iteable