+++ src/Language/Hasmtlib/Example/Arith.hs
@@ -5,7 +5,7 @@ import Language.Hasmtlib
 
 main :: IO ()
 main = do
-  res <- solveWith (solver cvc5) $ do
+  res <- solveWith (solver @SMT cvc5) $ do
     setLogic "QF_LIA"
 
     x <- var @IntSort
@@ -14,7 +14,7 @@ main = do
     assert $ y >? 0 && x /== y
     assert $ x `mod` 42 === y
     assert $ y + x + 1 >=? x + y
-    
+
     return (x,y)
 
   print res