+++ src/Language/Hasmtlib/Internal/Render.hs
@@ -2,9 +2,10 @@ module Language.Hasmtlib.Internal.Render where
 
 import Data.ByteString.Builder
 import Data.Foldable (foldl')
+import Data.Sequence
 import GHC.TypeNats
 
--- | Render values to their SMTLib2-Lisp form, represented as @Builder@.
+-- | Render values to their SMTLib2-Lisp form, represented as 'Builder'.
 class Render a where
   render :: a -> Builder
 
@@ -28,9 +29,17 @@ instance Render Double where
     | otherwise = formatDouble standardDefaultPrecision x
   {-# INLINEABLE render #-}
 
+instance Render Char where
+  render = char8
+  {-# INLINE render #-}
+
+instance Render String where
+  render = string8
+  {-# INLINE render #-}
+
 instance Render Builder where
   render = id
-  {-# INLINEABLE render #-}
+  {-# INLINE render #-}
 
 renderUnary :: Render a => Builder -> a -> Builder
 renderUnary op x = "(" <> op <> " " <> render x <> ")"
@@ -49,3 +58,7 @@ renderNary op xs = "(" <> op <> renderedXs <> ")"
   where
     renderedXs = foldl' (\s x -> s <> " " <> render x) mempty xs
 {-# INLINEABLE renderNary #-}
+
+-- | Render values to their sequential SMTLib2-Lisp form, represented as a 'Seq' 'Builder'.
+class RenderSeq a where
+  renderSeq :: a -> Seq Builder